Somatic Post Synaptic Current (Dpsc) calculation function
updated finding that K is a 'summation index in a symbolic sum' where a symbolic sum is such as sum(xi+yi) where x and y are symbols and i is the summation index.  I am still try trying to understand how can that be represented in matlab in a formula?