Neurons Total Soma Membrane Potential (neuronTSMP) Function Calculation
This implementation is not fully working now because I am having challenges with the formulas in totalSomaMembranePotential.  More is described in that function file but could anyone help me with that?  I will continue actively trying to get it fully working as well.  In this update I changed the function to work with the current code