Added logging off by default and included scoring function
Function was added that produces 3 scoring statistics based on the article's results analysis.  These are some basic scoring methods without statistical significance calculated for them.  Potentially such significance calculations can be added later.  Results are reported at selected epoch intervals.

uniqueSpikePercentage = the average number of spikes produced for each 
Char.  It represents the amount of neurons that uniquely identify with each 
Char.  Based on the article's analysis of which neurons fire specifically 
for each Char.

topVsClosestNtmpTotal = The difference in membrane potential between
the top potential and the one closest to it.  Represents how distinctly
the top responding neuron's potential was compared to the 2cnd highest.
Based on figure 7 in the article that shows the difference in potentials.

topVsAllNtmpTotal = The difference in membrane potential between
the top potential and the average of the others.  Represents how distinctly
the top responding neuron's potential was compared to the average of the others.
Based on figure 7 in the article that shows the difference in potentials.