Added additional numberOfSpikesPerChar scoring statistic
numberOfSpikesPerChar describes per Char what the number of spikes fired are for each epoch.  This helps describe how discriminating the neurons are in firing spikes.  In the future I will try to rebase the master to my branch and issue a pull request but this time I just wanted to add the code directly, I tested the latest version of this file with the updates and seen that it works.  Are there any other results reporting that there is interest in adding?  Feel free to modify any of the logging or results code and let me know if there are any requested changes I can work on.