Corrected reporting to use 100ms per epoch.  Changed totalMembraneMv to totalPositiveMembraneMv.
Incorrect timescale was found to be used previously in results reporting and that was corrected.  Also, in this Brian model it was found that membrane potential can be negative at times which creates challenges when reporting total membrane potential.  The total membrane potential was based on the result in figure 7 in the gupta and long paper: http://i.imgur.com/ZJAYptd.jpg .  To still capture some of the meaning in this model only the positive Mv are reported for this stat.  An issue is that in some neuron epochs that never fire a spike their total Mv is greater than those that fire spikes it seems to me due to no depolarization phase, that in mind the stat still has meaning but I am looking into a better one.