Work in progress toward adding formulas from gupta paper more exactly
I have added an initial effort to include part of the weight calculation used exactly in the paper.  The weight calculation is meant to run alongside the voltage membrane potential being calculated but seems to only have 2 iterations when included in the 'eqs' model for ADDS.  I expected it to run for each 4 neurons in each of the 10000ms and I will look into why it is not haivng more iterations than 2. I find that part of weight calculations need to be linked to spike firing times and therefore the weight calculations need to be synchronized with the membrane potentials being calculated and creating spikes.