Implemented further formulas in the Gupta paper according to the Tsmp.Calcs.ipynb computations.
This includes an implementation of the formulas for Id, Is, Um, Td, and Rd which is intended to be exactly as they are in the Gupta paper.  Currently the output seems to start out with creating different Um for each neuron for many ms but ends up becoming nan which I think means 0 from division from 0 or something.  I'll check why it is getting that result.  Output is displayed as command line standard output.  Also, I have not seen how to implemented it using the brian functions but I will work on that too.  Removed some commented out code too.